Use common header to drive generation.
qmlbind, at least in newer versions, ships with a single header that
includes the others in the project. SWIG can't process that file directly
(it doesn't recurse into #includes), but we can pre-process it a bit to
turn it into a valid binding file, and not have to worry about maintaining
the include list.